Merging Seiver with Master

+++ project1.py
@@ -375,8 +375,8 @@ def SearchAndDelete(c, conn, username):
         print("")
         print("1. Search for a request")
         if len(rows) != 0:
-            print("2. Delete a ride request\n")
-
+            print("2. Delete a ride request")
+        print("")
         choice = input("Please enter an option number, or 'exit': ")
         if choice == "1":
             code = (input("Please enter a location code, or city name: ")).capitalize()
@@ -400,8 +400,13 @@ def SearchAndDelete(c, conn, username):
 
         elif choice == "2" and len(rows) != 0:
             selection = int(input("Please enter the index number above of the request you would like to delete: "))
+            if selection < 1 or selection > len(rows):
+                print("Invalid index")
+                input("Press enter to continue...")
+                continue
             c.execute('''DELETE FROM requests WHERE
                          rid = ?''', (rows[selection - 1][0],))
+            conn.commit()
         elif choice == "exit":
             return 1
         else:
